backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
anthropic==0.86.0
langchain==1.2.13
langchain-anthropic==1.4.0
langchain-chroma==1.1.0
chromadb==1.5.5
pyyaml==6.0.3
python-dotenv==1.2.2
httpx==0.28.1
